Simplify ZygoteForkDelegate API further
This patch makes three changes:

  1. Removes the AckChild() delegate method used to send a custom
     message to the child process.
  2. Instead, the parent always writes the child's PID (as seen by the
     browser) over the pipe.  (Exception: When writing to a NaCl child
     process, we instead send 0 to avoid leaking the real PID into the
     NaCl address space.)
  3. Makes the Fork() delegate method responsible for sending the IPC
     channel ID to the child process.

This is in preparation for the next patch which will switch the pipe
direction to make the child responsible for discovering its own PID
and sending it to the parent process.  By removing AckChild(), this
simplifies the protocol and makes this change easier to implement.

BUG=357670

Review URL: https://codereview.chromium.org/240673002