GVDAH: Remove the NOTIFY_ERROR() macro.
The way it is right now (with a `;') will not work as
expected if used under a non-braced if-else. This is
actually already happening in GVDAH::Send().

We could fix that, but it's not providing us too much value
here, so get rid of it instead.

BUG=None
TEST=Compile and run

Review URL: https://codereview.chromium.org/242943003