Fix an issue that video surface view will block infobar
When entering fullscreen video, the surface view's size is set to the size of the screen if video size is unknown.
That blocks the display of infobar.
For EME video, however, the user has to acknowledge the infobar before the video starts to play.
And since the infobar is blocked, user cannot acknowledge the infobar. And thus video won't play.

BUG=364880

Review URL: https://codereview.chromium.org/243123003