Fix a bug that phone number field is not saved occasionally.
An editing phone number value is not applied correctly because a reply for a
chrome.send() has some delay.

Bug scenario:
1. Click handler for autofill-edit-address-apply-button is kicked.
2. document.activeElement.blur() is called.
 - An INPUT element in phone-list loses focus.
 - AutofillPhoneValuesList.validateAndSave is called.
 - It calls chrome.send('validatePhoneNumbers')
3. 0ms delay by setTimeout in the click handler
4. self.saveAddress_() is called, and the overlay is dismissed.
5. AutofillEditAddressOverlay.setValidatedPhoneNumbers is called.  It's too late.

To fix this issue, we wait for completing phone number validation.
 - AutofillPhoneValuesList provide a Promise which is fulfilled when all
   of validation requests are completed.
 - The autofill-edit-address-apply-button handler wait until the
    Promise is fulfilled.

This process worked well accidentally before r254991 because we incorrectly did
blur in the mousedown handler of autofill-edit-address-apply-button, and
setValidatedPhoneNumbers was called before mouseup.

BUG=361950

Review URL: https://codereview.chromium.org/236003008