Fix machine_model behaviors in gpu blacklist.
We made the assumption that GPUInfo.machine_model is (name, version).

This is true for mac, where this machine_model was originally designed for.
Now we also use it on Android. However, the whole machine_model should just
be retreated as one name string.

Now on Android Nexus, we have "Nexus 5", "Nexus 7", so they are processed
like on Mac, which is incorrect.

I believe a few android entries in the software_rendering_list_json.cc
are not active right now due to the (name, version).

BUG=364743
TEST=gpu_unittests
R=kbr@chromium.org, tsepez@chromium.org, vrk@chromium.org
TBR=pfeldman@chromium.org

Review URL: https://codereview.chromium.org/241793002