events: Fix event dispatch when the dispatcher list changes during dispatch.
Avoid crashing if the dispatcher list changes during dispatching an event. This
is done by using an ObserverList<> instead of a std::vector<> for the dispatcher
list. Update the doc for Add/RemovePlatformEventDispatcher() to clarify the
expected behaviour when dispatchers are added/removed during dispatching an
event.

BUG=360477
R=sky@chromium.org

Review URL: https://codereview.chromium.org/244293002