Introduce weak pointer for file_manager::VolumeManager (and some clean up.)
Although the current usages of Unretained(this) bound to callbacks are safe,
in the near future (for the associated BUGs) we'll need to pass callback to
external object that lives longer than the VolumeManager itself.

This patch introduces WeakPtrFactory, piggybacked with some tiny cleanup.

BUG=363960, 359418

Review URL: https://codereview.chromium.org/246963002