Mark WebViewClient subclasses that return a null layerTreeView.
In a followup blink-side patch, returning a null layerTreeView without
returning true from allowsBrokenNullLayerTreeView will assert.
Soon, we'll delete the null layerTreeView codepath entirely.
This is to make sure we have all the cases identified and so
that new ones don't get added.

Remove the null-check in RenderWidget::initializeLayerTreeView.
It never sets compositor_ to null and on the blink side we
assert as such.

Review URL: https://codereview.chromium.org/242953005