Password bubble: Remove IconView from BubbleView.
The BubbleView currently holds a reference to the IconView in order to
change its tooltip. We shouldn't do that; that should be taken care of
by migrating the bubble view to a BubbleIconView, which will handle
visibility changes and tooltop status. That will happen in a followup
CL.

BUG=261628

Review URL: https://codereview.chromium.org/230463002