Convert notifications to use RenderFrame instead of RenderView.
ContentBrowserClient::CancelDesktopNotification is removed in favor of having ShowDesktopNotification return a callback to cancel the notification, which seems simpler.

DesktopNotificationDelegate is a new interface passed to ShowDesktopNotification which hides the implementation details of sending IPCs to the right RenderFrame, and the notification_id parameter.

BUG=304341
R=nasko@chromium.org

Review URL: https://codereview.chromium.org/241553002