Fix an issue that fullscreen video will not resume after being interrupted by other apps
When video is interrupted, Release() is called on the MediaPlayerBridge.
In the old fullscreen mode, BrowserMediaPlayerManager::FullscreenPlayerPlay() should re-attach the surface to the new mediaplayer.
However, with the new fullscreen mode, FullscreenPlayerPlay() is not going to be called since there is no android media control.
Instead, hitting the play button will call BrowserMediaPlayerManager::OnStart().
As a result, we should copy the logic that re-attaching the surface into OnStart(). And we should not attach the embedded surface to the media player when video is in fullscreen.

BUG=365310

Review URL: https://codereview.chromium.org/243483006