Add the ability to better mock BrowserContextKeyedServices for testing.
This is particularly useful in browser tests: Currently, in these tests, a
testing factory can only be set once the TestingProfile is already constructed.
This can be quite inconvenient when trying to install test versions of classes
that the TestingProfile will instantiate by default, as some keyed services are
not designed to ever be destroyed other than when the BrowserContext is shut
down. See https://codereview.chromium.org/196783002/ for an example.

BUG=347247
TEST=none
R=erg@chromium.org

Review URL: https://codereview.chromium.org/244083002