Removes InputMethodContainer, merge into SettingsBubbleContainer (2)
We don't have to have a container for IME separately, the settings
bubble container is better for the IME UIs.

This switch also makes the following effects:
- IME UIs would appear in the lock screen / login screen
- IME UIs wouldn't move across the root window when removed (like
  display state switch among mirroring<->extended)

The first is good, because we can show the mode indicator when
switching the layout between qwerty and Dvorak. Candidate windows
will not appear in lock/login screen anyways.

The second seems no problem because IME UIs are more ephemeral
anyways. People won't get frustrated if disappears in such situation.
They will re-appear when the user types another key.

This is the reland of r265161. r265161 was reverted because
there's a conflict about keyboard focus among the candidate window
and the text-input context. This CL contains the fix for this
situation.

BUG=364667
TBR=oshima@chromium.org, shuchen@chromium.org

Review URL: https://codereview.chromium.org/248133004