cc: Use TaskGraphRunner instance in DirectRasterWorkerPool to run tasks.
This makes the DirectRasterWorkerPool code consistent with
ImageRasterWorkerPool. At the moment this means some code
duplication but that can be reduced in follow up patches.

This is expected to have a negative effect on perftests.
DirectRasterWorkerPool is expected to have similar
performance as ImageRasterWorkerPool after this change.

BUG=365840

Review URL: https://codereview.chromium.org/246633003