Remove tests for showModalDialog
This deletes the chromium-side tests for showModalDialog. We're about to disable
this feature from the Blink side, which would otherwise break these tests. The
plan is to leave the implementation behind a flag for a release or two in case
we need to restore the feature on short notice.

This CL is a partial re-land of https://codereview.chromium.org/165793003.

R=darin@chromium.org
BUG=345831

Review URL: https://codereview.chromium.org/246013006