Changes ViewManager to be node based and changes ids
ViewManager is now in terms of nodes. I'll add views next. I've also
changed ids so that they are uint16 and uint32. I'm a bit worried this
is going to be easy to get wrong (harder with a struct), but we'll try
this for now.

BUG=365012
TEST=covered by tests
R=ben@chromium.org

Review URL: https://codereview.chromium.org/246413007