Fix notification crash when a page removes it using notification.close().
The problem was introduced in r265295, because the string holding the id gets deallocated when the callback is removed from a map, but it used later. This is tested by NotificationsTest.TestCancelNotification, but the notification tests were accidentally removed in r262988, so readding them.

BUG=366090
R=stevenjb@chromium.org

Review URL: https://codereview.chromium.org/250043002