TRACING: Enable java tracing from native also
Currently java traces are enabled in Java only when tracing
that is triggered from an Android intent. This flips it
around so we always notify Java to start tracing, no
matter how tracing is enabled.

The case that wasn't working in this case is the 'dev tools
tracing handler' which is how tracing is enabled for
telemetry.

BUG=None

Review URL: https://codereview.chromium.org/237853005