x11: Do an XFlush after XSendEvent.
As djkurtz@ explains so well in https://codereview.chromium.org/249253005/,
X does an implicit flush of the output buffer when calling XPending. When
the glib-based message-pump is active, we do a XPending at every iteration
of the event-loop to see if there is events to be read. For this reason,
it is not necessary to do an explicit XFlush after XSendEvent. However, if
we switch to the libevent based message-pump, then we don't use XPending
to check whether or not there are pending events. So do an explicit XFlush
after XSendEvent.

BUG=240715
R=oshima@chromium.org

Review URL: https://codereview.chromium.org/254443002