Mac: Add tracing of the CoreAnimation presentation state
Track two pieces of state useful in debugging Mac drawing
performance: (1) the period over which a swap ack is pending
and (2) the instants at which CoreAnimation queries if we have
content to draw.

Ideally these would be trace instants and async events on a
pseudo-thread, but that functionality is incomplete. Instead,
use a counter to indicate the state, and instantaneous jumps
in the counter for the instants.

BUG=

Review URL: https://codereview.chromium.org/246933010