WebRTC recipe: Fix revision passing for Android APK commit bots.
When deploying these bots to the client.webrtc build
waterfall, revision information is passed to the gclient sync
command like this example: --revision src@5919
This happens because the code that handles overwriting that
was only turned on for the trybot part of the recipe.

This CL fixes this and also makes use of the 'webrtc_revision'
variable which produces a less hacky gclient spec.

BUG=346198
TEST=I ran:
scripts/tools/run_recipe.py webrtc/android_apk buildername="Android Chromium-APK Builder" mastername="client.webrtc" revision="'5956'"
and verified the right revision flag gets passed and that the build is successful.

Review URL: https://codereview.chromium.org/247943002