Make hidden app windows have no shelf presence [ChromeOS].
This CL fixes an issue with hidden app windows appearing in the shelf by
unregistering the app windows when they are hidden and registering them
again when shown.

Window visibility wasn't enough here as the window is not visible
if it is minimized.

This CL adds OnAppWindowShown() and OnAppWindowHidden() methods to
AppWindowObserver.

BUG=360896

Review URL: https://codereview.chromium.org/234673002