Include wifi_component only when in Win or Mac platforms.
This issue was found while doing
https://codereview.chromium.org/243343002 and we discovered that android
was building wifi_test target.

It was building this target because it does:

$ ninja -C out/Debug -j50

And that builds all the targets generated by gyp.

Since the target was not excluded from android platform, a target was
created there and thus the above command would build it.

This patch fixes this problem, by only including it where it is actually
supported.

BUG=None
TEST=green trybots
R=mef@chromium.org,thestig@chromium.org
TBR=jochen@chromium.org

Review URL: https://codereview.chromium.org/259543002