Remove RenderWidgetHost::GetSnapshotFromRenderer.
This functionality does not exist on aura or android anymore, and is
going away soon on Mac. RWHViews are always composited now on all
platforms, so the active tab is always able to be copied in the
browser process with CopyFromBackingStore().

This removes the ViewMsg_Snapshot and ViewHostMsg_Snapshot IPCs.

R=piman, thestig
BUG=251966, 251960

Review URL: https://codereview.chromium.org/259523008