Don't clear render surfaces unnecessarily.
As an optimization, we used to clear render surfaces when they were
elided from the RenderSurfaceLayerList. It turns out that these surfaces
are still important from time to time; a scroll child will need to
access the transforms of these surfaces in order to put its clip in the
right space.

This CL retains the layer lists and it also updates RenderSurface to
own a LayerList rather than a RenderSurfaceLayerList (which was always
a bit of a kludge).

R=danakj@chromium.org
BUG=366386

Review URL: https://codereview.chromium.org/250803013