No media-file validation mode for DeviceMediaAsyncUtil.
CreateSnapshotFile and GetFileStreamReader used to sniff the content
of the files and returned error if it does not look like media data.
This CL adds a mode to disable the check, and use the disabled mode
in MTP mounting in Chrome OS file manager.

BUG=367004

Review URL: https://codereview.chromium.org/251723004