Remove changes remaining counter from about:sync
The label "changes remaining" is not quite correct.  It's actually set
to either zero or one, depending on whether or not the current sync
cycle should result in more downloads.

If there is an issue with infinite download looping, we should be able
to detect and debug it based on the protocol logs, among other things.
The protocol logs indicate when a GetUpdates response had a non-zero
"changes remaining" value.

BUG=349301

Review URL: https://codereview.chromium.org/256413007