Indentation fix so trybots don't clobber twice on non-lkcr compile failure.
Theres a bug where if the first compile (with patch) fails, and the second compile
passes, the recipe compiles again anyways.  This should fix it, and since I noticed that
none of the expectations changed when it was fixed, this also adds a new regression
test.

BUG=361346

Review URL: https://codereview.chromium.org/260043006