Add more phony rules to GN build.
For a target //foo/bar:bar this adds the additional phony rules: "foo/bar:bar" and "foo/bar" to build.ninja. This is on top of the phony rules "bar" which will exist for all targets with a unique name "bar". This allows one to disambiguate targets with different paths but the same short name.

R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/247663006