Add ScopedStubEnterpriseInstallAttributes for tests to set test install attributes.
This change is motivated by the need to set the test install attributes before the browser policy connector is initialized for the first time. In particular, test class members may cause the policy connector to be initialized in their constructors, so we also need to set the install attributes in the constructor of a member.

BUG=243341

Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=266094

Review URL: https://codereview.chromium.org/247283007