cc: prevent TileManager raster/eviction tile iterators copy.
This patch ensures that iterators cannot be copied..
Since we rely on a member heap that points to another member vector,
we can't do a naive copy. Instead we would have to reconstruct the
heap to point to the new members, but it's better to just restrict
copy altogether.

R=enne

Review URL: https://codereview.chromium.org/251003002