Reliably name temporary uploads/downloads by their buildnumber.
Instrumentation bot was failing because the builder was uploading a build product based on a revision, and the tester was downloading a build product based on the buildnumber.  I'm switching them all to buildnumber (for compatibility with try job when revision doesn't make sense).

Goes with https://chromereviews.googleplex.com/44657024 on internal side.

BUG=366447

Review URL: https://codereview.chromium.org/260123006