No need to delete ax subtree when new node is added.
AXTreeSerializer was instructing its client to delete the whole
subtree whenever it got a notification on a new node, but that
isn't actually necessary. It was causing the whole tree to be
recreated unneccessarily. Existing unit tests still pass, and
they're comprehensive.

BUG=368295
NOTRY=true

Review URL: https://codereview.chromium.org/260463003