Fix flickering Mac app launcher progress bar on <10.9
On Mavericks, the app lanucher subviews need to have access to a
compositing layer they can share. Otherwise the compositor makes tearing
artifacts. However, doing this on Mountain Lion or earlier results in
flickering whilst an item is installing IF the app launcher menu is also
opened - a different Apple bug.

This CL restricts the fix for the tearing to Mavericks or later.

BUG=362547
TEST=Install a large app from the webstore on Mountain Lion (10.8) or
earlier. Open the app launcher menu. The progress bar should not
flicker.

Review URL: https://codereview.chromium.org/263613002