Use chrome.app.window full-screen API for apps v2.
This allows us to hook onMaximized to enter full-screen which is a cleaner UX.
In the future, we will also have on-screen auto-hide close/minimize/restore
controls in full-screen mode, which will eliminate the need for the tool-bar
menu entry completely.

See https://developer.chrome.com/apps/app_window#type-AppWindow for details
of the API, and http://code.google.com/p/chromium/issues/detail?id=364942 for a discussion of why
this behaviour is not going to be supported by requestFullscreen.

BUG=134213,252927
NOTRY=true

Review URL: https://codereview.chromium.org/252783003