Make PID translation sanity checks more robust
Instead of checking that the PID the kernel told us for the zygote
happens to be one of our descendents, we have the zygote send us an
extra message right away.  This gives us a check of PID translation
without any false positives or negatives.

BUG=357670

Review URL: https://codereview.chromium.org/254093003