Trusted plugin: Clean up manifest error reporting.
This streamlines some of the error reporting in the Manifest support classes in
the trusted plugin. Without this change, it's a bit hard to tell when we report
errors and when we don't; this makes some logic more explicit for when errors
are reported.

This will hopefully make reviewing the change that moves the manifest support
out of the trusted plugin easier.

BUG=353592

Review URL: https://codereview.chromium.org/266723004