Prevent extensions from calling DownloadTargetDeterminer::DoLoop the wrong number of times.
Clear the callbacks from ExtensionDownloadsEventRouterData when one is run so that it won't be re-run when the event page is unloaded.

Add DCHECKs to callbacks in DownloadTargetDeterminer to check that they are only called when they are expected.

Add a DelayedTask timeout to make sure that a callback is called within 15 seconds.

BUG=360332,302071

Review URL: https://codereview.chromium.org/249853002