Add _api_features.json for chrome.shell API
The chrome.shell API is only available in app_shell, so it should not be part
of the common extensions _api_features.json.

This fixes a CHECK failure in the renderer when loading the Calculator sample
app, which tests for the existence of chrome.shell.  (The schema is not
available, because the API is only implemented in app_shell, but chrome.shell
returns true because the "shell" is in the shared extensions _api_features.json.)

BUG=369318
TEST=manual

Review URL: https://codereview.chromium.org/260963006