Add TestPasswordStore::IsEmpty()
There was no simple way to test the emptiness of the TestPasswordStore.
That made it easy to test the emptiness in a wrong way, see http://crbug.com/368778#c8.
The problem was the assumption that
password_store.saved_passwords().size() == 0u
is equivalent with no passwords being saved.

In fact, only the => implication holds true, the <= not -- it can happen that the map saved_passwords() has an entry, which consists of an empty vector of passwords. If, and only if, all entries are like that, the password store contains no passwords.

This CL introduces an IsEmpty method for the TestPasswordStore, checking the above condition for emptiness.
It also cleans-up the signature of a related accessor saved_passwords().

BUG=368778, 367690
R=rchtara@chromium.org,mkwst@chromium.org

Review URL: https://codereview.chromium.org/267803004