Properly reset dirty bits before early-out in PreCalculateMetaInformation.
We can early-out in this function if we have a non-invertible, non-animated
transform. This causes us to not reset some dirty bits (in particular,
|sorted_for_recursion|). Later, when sorting a layer's child list (_before_
we decide to early-out of CalculateDrawProperties), we skip out on sorting
the layer due to the dirty state. This causes badness. We should reset the
state and sort the list properly, including the layers that will be skipped
later.

R=vollick@chromium.org,ajuma@chromium.org,danakj@chromium.org
BUG=366233

Review URL: https://codereview.chromium.org/267713011