ChromeOS: Avoid overlapping app list with virtual keyboard.
When the app list is centered (--enable-experimental-app-list-position),
and the virtual keyboard is active, the app list is centered in the
screen rect that excludes the keyboard region.

BUG=363731
TEST=Run with --enable-experimental-app-list-position
--enable-virtual-keyboard. Open the app list, then the virtual keyboard.
The app list should move upwards so it is centered in the part of the
screen without the virtual keyboard.
TEST=Same as above, but with the additional flag
--enable-virtual-keyboard-overscroll.

Review URL: https://codereview.chromium.org/250423004