Add get_target_outputs function to GN
This function returns the outputs for a target that appeared previously in the same file. This is frequently needed for getting the files resulting from an action.

The main operational change is that targets that are generated are stashed in a vector as we process a file and then dispatched once that file is finished, rather than being dispatched as-we-go. This way we can ask questions about the targets that appeared previously in the file.

R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/269723006